+++ backend/files/file.go
@@ -206,6 +206,7 @@ func MoveResource(source, realsrc, realdst string, isSrcDir bool) error {
 	if err != nil {
 		return err
 	}
+
 	index := GetIndex(source)
 	// refresh info for source and dest
 	err = index.RefreshFileInfo(FileOptions{